requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
arq==0.26.1
redis[hiredis]==5.2.1
pyjwt==2.10.1
httpx==0.28.1
bcrypt==4.2.1
pydantic-settings==2.7.0
lxml==5.3.0
geoalchemy2==0.15.2
aiosqlite==0.20.0
